impl ExposedUserFinder

+++ user-service/database/changelog/0.0.1/01_init_table.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS users.user(
-    id INT PRIMARY KEY NOT NULL,
+    id BIGINT PRIMARY KEY NOT NULL,
     first_name VARCHAR(50) NOT NULL,
     last_name VARCHAR(50) NOT NULL
 );
